EC Signed DH: recover step 0
A quick note: the definition of SignedDH might
be better expressed by recomputing g^esk when
responding (so session state only needs to keep
the ephemeral secret).
This would (sooner or later) impose a
functional-style invariant ('in every session's
state st, we have st.epk = g^st.esk'; because the
reduction will anyway use epk!), of a kind that is
not always fun. So we don't do that. But we could.